@@ -15,7 +15,7 @@ classifiers = [
   'Programming Language :: Python :: 3.10',
   'Programming Language :: Python :: 3.11',
   'Programming Language :: Python :: 3.12',
-  'Programming Language :: Python :: 3.9',
+  'Programming Language :: Python :: 3.13',
   'Topic :: Scientific/Engineering',
   'Topic :: Software Development',
 ]
@@ -42,7 +42,7 @@ maintainers = [
 ]
 name = 'mne-icalabel'
 readme = 'README.md'
-requires-python = '~=3.9'
+requires-python = '~=3.10'
 version = "0.8.0.dev0"
 
 [project.optional-dependencies]
@@ -171,7 +171,6 @@ minversion = '8.0'
 [tool.ruff]
 extend-exclude = []
 line-length = 88
-target-version = 'py39'
 
 [tool.ruff.format]
 docstring-code-format = true
@@ -186,8 +185,6 @@ select = ['A', 'B', 'D', 'E', 'F', 'G', 'I', 'LOG', 'NPY', 'PIE', 'PT', 'T20', '
   'D100', # 'Missing docstring in public module'
   'D104', # 'Missing docstring in public package'
   'D107', # 'Missing docstring in __init__'
-  'UP007', # 'Use `X | Y` for type annotations', requires python 3.10
-  'UP038', # 'Use X | Y in {} call instead of (X, Y)', requires python 3.10
 ]
 '*.pyi' = ['E501']
 '__init__.py' = ['F401']
