@@ -13,7 +13,8 @@ from typing import TYPE_CHECKING
 from mne.utils.docs import docdict as docdict_mne
 
 if TYPE_CHECKING:
-    from typing import Any, Callable
+    from collections.abc import Callable
+    from typing import Any
 
 # -- Documentation dictionary ----------------------------------------------------------
 docdict: dict[str, str] = {}
